Add files via upload
This program is a ML model to predict the chance of approval of the loan offer, based on supervised classification. 
Different algorithms are developed including KNN, LogReg, and Naive Bayes, to study their accuracy in prediction. 
data processing of zip codes is performed. 
correlation and data visualization is provided. 
method of Cross Validation is devloped. 
Models are evaluated.